+++ components/vf-analytics-google/CHANGELOG.md
@@ -1,6 +1,8 @@
 ### 1.0.0-rc.3
 
 * documentation cleanup
+* `analyticsTrackInteraction()` is now namespaced as `vfGaTrackInteraction()`
+* `vfGaTrackInteraction()` now documented for direct usage
 
 ### 1.0.0-rc.2